add migration

+++ src/tasks/tasks.controller.ts
@@ -3,6 +3,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -22,6 +23,7 @@ import { CreateTaskLabelDto } from './create-task-label.dto';
 import { FindTaskParams } from './find-task.params';
 import { PaginationParams } from 'src/common/pagination.params';
 import { PaginationResponse } from 'src/common/pagination.response';
+import { CurrentUserId } from 'src/users/decorators/current-user-id.decorator';
 
 @Controller('tasks')
 export class TasksController {
@@ -31,38 +33,53 @@ export class TasksController {
   public async findAllTasks(
     @Query() filters: FindTaskParams,
     @Query() pagination: PaginationParams,
+    @CurrentUserId() userId: string,
   ): Promise<PaginationResponse<Task>> {
     const [tasks, total] = await this.tasksService.getAllTasks(
       filters,
       pagination,
+      userId,
     );
     return {
       data: tasks,
       meta: {
         total: total,
-        offset: pagination.offset,
-        limit: pagination.limit,
+        ...pagination,
+        // offset: pagination.offset,
+        // limit: pagination.limit,
       },
     };
   }
 
   @Get('/:id')
-  public async findOneTask(@Param() params: FindOneParams): Promise<Task> {
-    return await this.findOneOrFailed(params.id);
+  public async findOneTask(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    const task = await this.findOneOrFailed(params.id);
+    this.checkTaskOwnership(task, userId);
+    return task;
   }
 
   @Post()
-  public async createOneTask(@Body() createTaskDto: CreateTaskDto) {
-    console.log('createTaskDto:', createTaskDto);
-    return await this.tasksService.createOneTask(createTaskDto);
+  public async createOneTask(
+    @Body() createTaskDto: CreateTaskDto,
+    @CurrentUserId() userId: string,
+  ) {
+    return await this.tasksService.createOneTask({
+      ...createTaskDto,
+      userId,
+    });
   }
 
   @Patch('/:id')
   public async updateTask(
     @Param() param: FindOneParams,
     @Body() updateTaskDto: UpdateTaskDto,
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFailed(param.id);
+    this.checkTaskOwnership(task, userId);
     try {
       return this.tasksService.updateTask(task, updateTaskDto);
     } catch (error) {
@@ -76,8 +93,12 @@ export class TasksController {
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  public async deleteOneTask(@Param() param: FindOneParams) {
+  public async deleteOneTask(
+    @Param() param: FindOneParams,
+    @CurrentUserId() userId: string,
+  ) {
     const task = await this.findOneOrFailed(param.id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.deleteOneTask(task);
   }
 
@@ -85,9 +106,11 @@ export class TasksController {
   public async addLabels(
     @Param() param: FindOneParams,
     @Body() labels: CreateTaskLabelDto[],
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     console.log('addLabels -> labels:', labels);
     const task = await this.findOneOrFailed(param.id);
+    this.checkTaskOwnership(task, userId);
     return await this.tasksService.addLabels(task, labels);
   }
 
@@ -106,8 +129,16 @@ export class TasksController {
   public async deleteLabels(
     @Param() param: FindOneParams,
     @Body() deleteLabels: string[],
+    @CurrentUserId() userId: string,
   ): Promise<void> {
     const task = await this.findOneOrFailed(param.id);
+    this.checkTaskOwnership(task, userId);
     this.tasksService.deleteTaskLabels(task, deleteLabels);
+  }
+
+  private checkTaskOwnership(task: Task, userId: string): void {
+    if (task.id !== userId) {
+      throw new ForbiddenException('You can only access your own tasks');
+    }
   }
 }
